int64 to int

+++ Ejercicio02_Listas/main.go
@@ -5,7 +5,7 @@ import (
 )
 
 func main() {
-	ListaDePares, ListaDeImpares, ListaDeSuma := OrdenarYSumar([]int64{12, 5, 12, 2, 33, 67, 99, 44, 3, 4, 67, 321, 9})
+	ListaDePares, ListaDeImpares, ListaDeSuma := OrdenarYSumar([]int{12, 5, 12, 2, 33, 67, 99, 44, 3, 4, 67, 321, 9})
 	fmt.Println("Lista De Pares", ListaDePares)
 	fmt.Println("Lista De Impares", ListaDeImpares)
 	fmt.Println("Lista De Suma", ListaDeSuma)
@@ -20,9 +20,9 @@ func main() {
 		input: [12,21,33,4] -> output: [4,12], [21,33], [25,35]
 		input: [12,21,33,4,1] -> output: [4,14], [1,21,33], [5,35,33]
 */
-func OrdenarYSumar(numeros []int64) ([]int64, []int64, []int64) {
+func OrdenarYSumar(numeros []int) ([]int, []int, []int) {
 
 	// SOLUCIÓN
 
-	return []int64{}, []int64{}, []int64{}
+	return []int{}, []int{}, []int{}
 }
